refactor: improve code organization and reduce duplication
This commit makes several improvements to the codebase structure:

- Consolidate module files (oauth2, passkey, session) into single files
- Split storage implementations into database-specific files (postgres.rs, sqlite.rs)
- Move common utility functions to centralized locations
- Fix static file path references
- Improve function visibility with appropriate pub(crate) and pub(super) modifiers
- Remove duplicated code for session ID extraction and token management
- Fix Duration type usage in HTTP client configuration
- Add appropriate error types and conversions
- Add note about Android compatibility issue with signalCurrentUserDetails

These changes maintain the same functionality while making the code more
maintainable and reducing duplication, following the principle of having
a single source of truth for each piece of functionality.
