+++ Readme.md
@@ -8,6 +8,7 @@
 - Schema check when initializing database connection. Make sure the schema the program is expecting is the same as the one in the database.
 - When using Delete User button, deletion of passkey credentials aren't notified to Passkey Authenticator.
 - Adjust visibility of functions, structs, enums, etc. What needs to be public?
+- signalCurrentUserDetails seems not working on Android device.
 
 ## Half Done
