--- libaxum/src/oauth2/mod.rs
@@ -1,4 +1,0 @@
-// mod extractor;
-mod handlers;
-
-pub use handlers::router;
