--- libaxum/src/passkey/mod.rs
@@ -1,4 +1,0 @@
-mod handlers;
-mod routes;
-
-pub use routes::{passkey_well_known_router, router};
