--- libaxum/src/session/mod.rs
@@ -1,3 +1,0 @@
-mod extractor;
-
-pub use extractor::AuthUser;
