+++ oauth2_passkey/src/passkey/mod.rs
@@ -1,11 +1,9 @@
-mod common;
 mod config;
 mod errors;
 mod main;
 mod storage;
 mod types;
 
-pub use common::init;
 pub use config::PASSKEY_ROUTE_PREFIX; // Required for route configuration
 pub use errors::PasskeyError;
 
@@ -17,3 +15,16 @@ pub use main::{
 
 pub use storage::PasskeyStore;
 pub use types::{CredentialSearchField, StoredCredential};
+
+pub async fn init() -> Result<(), PasskeyError> {
+    // Validate required environment variables early
+    let _ = *config::PASSKEY_RP_ID;
+
+    crate::storage::init()
+        .await
+        .map_err(|e| PasskeyError::Storage(e.to_string()))?;
+
+    PasskeyStore::init().await?;
+
+    Ok(())
+}
