+++ oauth2_passkey/src/passkey/storage/mod.rs
@@ -1,3 +1,5 @@
-mod passkey_store;
+mod postgres;
+mod sqlite;
+mod store_type;
 
-pub use passkey_store::PasskeyStore;
+pub use store_type::PasskeyStore;
