--- oauth2_passkey/src/session/common.rs
@@ -1,19 +1,0 @@
-use crate::session::{errors::SessionError, types::StoredSession};
-
-use crate::storage::CacheData;
-
-impl From<StoredSession> for CacheData {
-    fn from(data: StoredSession) -> Self {
-        Self {
-            value: serde_json::to_string(&data).expect("Failed to serialize StoredSession"),
-        }
-    }
-}
-
-impl TryFrom<CacheData> for StoredSession {
-    type Error = SessionError;
-
-    fn try_from(data: CacheData) -> Result<Self, Self::Error> {
-        serde_json::from_str(&data.value).map_err(|e| SessionError::Storage(e.to_string()))
-    }
-}
